+++ key-manager-test/src/test/scala/org/bitcoins/keymanager/util/HdUtilTest.scala
@@ -16,7 +16,7 @@ import org.bitcoins.core.crypto.ExtKeyVersion.{
   SegWitTestNet3Pub
 }
 import org.bitcoins.core.hd.{HDCoinType, HDPurpose, HDPurposes}
-import org.bitcoins.keymanager.KeyManagerUnitTest
+import org.bitcoins.testkit.keymanager.KeyManagerUnitTest
 
 class HdUtilTest extends KeyManagerUnitTest {
