+++ testkit/src/main/scala/org/bitcoins/testkit/keymanager/KeyManagerTestUtil.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.keymanager
+package org.bitcoins.testkit.keymanager
 
 import java.nio.file.Path
 
@@ -6,6 +6,7 @@ import org.bitcoins.core.config.Networks
 import org.bitcoins.core.crypto.AesPassword
 import org.bitcoins.core.hd.HDPurposes
 import org.bitcoins.keymanager.bip39.BIP39KeyManager
+import org.bitcoins.keymanager.{KeyManagerParams, WalletStorage}
 import org.bitcoins.testkit.BitcoinSTestAppConfig
 import org.bitcoins.testkit.core.gen.CryptoGenerators
 import org.scalacheck.Gen
