+++ website/sidebars.json
@@ -37,6 +37,9 @@
       "rpc/rpc-eclair",
       "rpc/rpc-bitcoind"
     ],
+    "Testkit": [
+      "testkit/testkit"
+    ],
     "Contributing": ["contributing", "contributing-website"],
     "Security": ["security"]
   }
